backend/requirements.txt
# ── Web framework ─────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# ── Database ─────────────────────────────────────────────────
motor==3.5.1
pymongo==4.8.0

# ── Validation / Serialization ───────────────────────────────
pydantic==2.12.5
pydantic[email]==2.12.5
email-validator==2.2.0

# ── Auth / Security ──────────────────────────────────────────
PyJWT==2.9.0
bcrypt==4.2.0
cryptography==46.0.5
python-dotenv==1.0.1

# ── PDF / Document Parsing ───────────────────────────────────
pdfplumber==0.11.8
PyPDF2==3.0.1
python-docx==1.1.2
pdfminer.six==20251107
pypdfium2==5.4.0

# ── ML: Embeddings & Vector Search ───────────────────────────
sentence-transformers==5.1.2
faiss-cpu==1.13.0
numpy==2.0.2
scikit-learn==1.6.1
scipy==1.13.1
torch==2.8.0
transformers==4.57.6

# ── LLM Client ───────────────────────────────────────────────
ollama==0.6.1

# ── Utilities ────────────────────────────────────────────────
aiohttp==3.10.5
requests==2.32.5
tqdm==4.67.3